requirements.txt
python-binance
python-dotenv
argparse
tabulate